requirements-dev.txt
asttokens==3.0.0
black==25.9.0
click==8.3.0
contourpy==1.3.3
cycler==0.12.1
decorator==5.2.1
emoji==2.14.1
executing==2.2.1
fonttools==4.60.1
ipython==9.5.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
kiwisolver==1.4.9
matplotlib-inline==0.1.7
mypy==1.18.2
mypy_extensions==1.1.0
packaging==25.0
parso==0.8.5
pathspec==0.12.1
pexpect==4.9.0
platformdirs==4.4.0
prompt_toolkit==3.0.52
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.2
pyparsing==3.2.5
pytokens==0.1.10
six==1.17.0
stack-data==0.6.3
traitlets==5.14.3
typing_extensions==4.15.0
wcwidth==0.2.13
